Conflicts:
	OPCApp/OPCApp.AuthManage/Views/UserListWindow.xaml

+++ OPCApp/OPCApp.AuthManage/ViewModels/UserListWindowViewModel.cs
@@ -10,6 +10,7 @@ using OPCApp.Domain.Models;
 using OPCApp.Infrastructure;
 using OPCApp.Infrastructure.DataService;
 using OPCApp.Infrastructure.Mvvm;
+using OPCApp.Infrastructure.Mvvm.Model;
 
 namespace OPCApp.AuthManage.ViewModels
 {
@@ -18,7 +19,7 @@ namespace OPCApp.AuthManage.ViewModels
     public class UserListWindowViewModel : BaseListViewModel<OPC_AuthUser>
     {
         private OPC_AuthUser _curUser;
-        public PageResult _prResult;
+        public PageDataResult<OPC_AuthUser> _prResult;
 
         public UserListWindowViewModel()
             : base("UserListWindow")
@@ -36,7 +37,7 @@ namespace OPCApp.AuthManage.ViewModels
         public NodeInfo NodeInfo { get; private set; }
         public ReadOnlyCollection<DelegateCommand> Commands { get; private set; }
 
-        public PageResult PrResult
+        public PageDataResult<OPC_AuthUser> PrResult
         {
             get { return _prResult; }
             set { SetProperty(ref _prResult, value); }
@@ -60,9 +61,9 @@ namespace OPCApp.AuthManage.ViewModels
         /*导出用户*/
         public DelegateCommand ExportUserCommand { get; set; }
         /*双击用户列表*/
-        public DelegateCommand DBGridClickCommand { get; set; }
-        public DelegateCommand UpdateUserCommand { get; set; }
-
+        public DelegateCommand DelUserCommand { get; set; }
+        public DelegateCommand RePasswordCommand { get; set; }
+   
         public int PageSize { get; set; }
         public int PageIndex { get; set; }
 
@@ -86,6 +87,11 @@ namespace OPCApp.AuthManage.ViewModels
         {
             if (CheckSelection())
             {
+                if (NodeInfo.SelectedNode.OrgId == "100")
+                {
+                    MessageBox.Show("父级组织机构节点不能删除", "提示");
+                    return;
+                }
                 GetOperationNode().DeleteOrg();
             }
         }
@@ -95,6 +101,7 @@ namespace OPCApp.AuthManage.ViewModels
             AddOrgCommand = new DelegateCommand(AddOrg);
             EditOrgCommand = new DelegateCommand(EditOrg);
             DeleteOrgCommand = new DelegateCommand(DeleteOrg);
+
             List<OPC_OrgInfo> orgList = AppEx.Container.GetInstance<IOrgService>().Search().ToList();
             Commands = new ReadOnlyCollection<DelegateCommand>(new[]
             {
@@ -129,7 +136,7 @@ namespace OPCApp.AuthManage.ViewModels
 
         private bool CheckSelection()
         {
-            if (NodeInfo.SelectedNode == null)
+            if (NodeInfo.SelectedNode==null)//建议改后台
             {
                 MessageBox.Show("请选择组织机构节点", "提示");
                 return false;
@@ -176,13 +183,21 @@ namespace OPCApp.AuthManage.ViewModels
             {
                 return;
             }
-            PrResult = new PageResult();
+            PrResult = new PageDataResult<OPC_AuthUser>();
             PrResult.Models = PrResultTemp.Result.ToList();
             PrResult.Total = PrResultTemp.TotalCount;
         }
-
-        private void DBGridClick()
+        /*重写 在选择组织结构 才能进行用户增加操作*/
+         public override bool BeforeAdd(OPC_AuthUser t)
         {
+            var curNode = GetOperationNode();
+            if (curNode==null||curNode.OrgId==null)
+            {
+                MessageBox.Show("请选择部门", "提示");
+                return false;
+            }
+             t.OrgId = curNode.OrgId;
+             return true;
         }
 
         /*初始化页面固有的数据值*/
@@ -194,17 +209,36 @@ namespace OPCApp.AuthManage.ViewModels
             SelectedFiledValue = "";
             SelectedFiled = "";
             SetStopUserCommand = new DelegateCommand(SetStopUser);
-            DBGridClickCommand = new DelegateCommand(DBGridClick);
+            DelUserCommand=new DelegateCommand(DelUser);
             PageIndex = 1;
             PageSize = 10;
-            PrResult = new PageResult();
+            PrResult = new PageDataResult<OPC_AuthUser>();
             CurModel = new OPC_AuthUser();
-            UpdateUserCommand = new DelegateCommand(UpdateUser);
+            RePasswordCommand = new DelegateCommand(RePassword);
         }
 
-        private void UpdateUser()
+        private void RePassword()
         {
-            EditAction(CurModel);
+            OPC_AuthUser user = CurModel;
+            if (user == null)
+            {
+                MessageBox.Show("请选择要操作的用户", "提示");
+                return;
+            }
+          var resultMsg=  AppEx.Container.GetInstance<IAuthenticateService>().ResetPassword(user);
+          MessageBox.Show(resultMsg.IsSuccess ? "重置密码成功" : "操作失败", "提示");
+        }
+
+        public void DelUser()
+        {
+            OPC_AuthUser user = CurModel;
+            if (user == null)
+            {
+                MessageBox.Show("请选择要操作的用户", "提示");
+                return;
+            }
+            DeleteAction(user);
+            SearchAction();
         }
 
         private void SetStopUser()
@@ -214,11 +248,10 @@ namespace OPCApp.AuthManage.ViewModels
             {
                 MessageBox.Show("请选择要操作的用户", "提示");
                 return;
-                ;
             }
             var iauth = GetDataService() as IAuthenticateService;
-            bool isValid = user.IsValid == true ? true : false;
-            iauth.SetIsStop(user.Id, isValid);
+            iauth.SetIsStop(user);
+            SearchAction();
         }
 
         protected override IBaseDataService<OPC_AuthUser> GetDataService()
@@ -226,27 +259,6 @@ namespace OPCApp.AuthManage.ViewModels
             return AppEx.Container.GetInstance<IAuthenticateService>();
         }
 
-        public class PageResult : BindableBase
-        {
-            private List<OPC_AuthUser> _models;
-            public int _total;
-
-            public PageResult()
-            {
-                Models = new List<OPC_AuthUser>();
-            }
-
-            public int Total
-            {
-                get { return _total; }
-                set { SetProperty(ref _total, value); }
-            }
-
-            public List<OPC_AuthUser> Models
-            {
-                get { return _models; }
-                set { SetProperty(ref _models, value); }
-            }
-        }
+     
     }
 }
